+++ plone/supermodel/utils.py
@@ -2,7 +2,10 @@ import os.path
 import sys
 import re
 
-from elementtree import ElementTree
+try:
+    from lxml import etree as ElementTree
+except ImportError:
+    from elementtree import ElementTree
 
 from zope.interface import directlyProvidedBy, directlyProvides
 from zope.schema.interfaces import IField, IFromUnicode, IDict, ICollection